Issue #11163: Enforce file size on CovariantEqualsCheck
