+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/covariantequals/InputCovariantEqualsone.java
@@ -10,7 +10,7 @@ package com.puppycrawl.tools.checkstyle.checks.coding.covariantequals;
  * Test file for covariant equals methods.
  * @author Rick Giles
  */
-public class InputCovariantEquals
+public class InputCovariantEqualsone
 {
     private class Inner
     {
@@ -33,7 +33,7 @@ public class InputCovariantEquals
         }
     }
 
-    public boolean equals(InputCovariantEquals aInputCovariantEquals) // violation
+    public boolean equals(InputCovariantEqualsone aInputCovariantEquals) // violation
     {
         return false;
     }
@@ -108,40 +108,3 @@ interface InputCovariant6
     public boolean equals(InputCovariant5 aInputCovariant5);
 }
 
-class InputGenericCovariant7
-{
-    public <A> boolean equals(InputGenericCovariant7 aInputCovariant7)
-    {
-        return true;
-    }
-
-    public boolean equals(Object aObject)
-    {
-        return false;
-    }
-}
-
-class InputGenericCovariant8
-{
-    public <A> boolean equals(InputGenericCovariant8 aInputCovariant8)
-    {
-        return true;
-    }
-
-    public boolean equals(Object aObject)
-    {
-        return false;
-    }
-}
-
-enum InputEnumCovariant {
-    EQUALS;
-
-    public boolean equals(InputEnumCovariant inputEnumCovariant) { // violation
-        return false;
-    }
-
-    int equals(Integer integer) { // violation
-        return 0;
-    }
-}
